requirements/base.txt
# Base
pytz
python-slugify
jinja2

openai
pydantic

# Django
Django==3.2
django-cors-headers==3.7.0
django-rest-passwordreset
django-extensions
django-secured-fields

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.4
django-filter==2.4.0
drf-nested-routers==0.93.4
djangorestframework-simplejwt==4.6.0
pyJWT

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.2.0

# AWS
boto3==1.17.64

# Storage
psycopg2-binary==2.9.3
django-storages

# Channels
channels
channels_redis


# CMS
django-modeltranslation

# Images
pillow

# Swagger
drf-yasg==1.21.4

# Authentication
twilio

# Cipher
django-cryptography==1.1

# interface
django-admin-interface

# Google Cloud
google-cloud-storage
google-cloud-bigquery
google-cloud-firestore
google-cloud-datacatalog
google-cloud-dataplex
google-cloud-dataproc
google-cloud-iam
google-cloud-resource-manager
grpc-google-iam-v1
google-api-python-client

# MIME TYPES
python-magic

# Dataframes
pandas

#Notebooks
google-cloud-notebooks

